refactor: 구성 요소들의 네이밍 개선
