+++ src/main/java/com/couponpop/security/blacklist/repository/InMemoryTokenBlacklistRepository.java
@@ -10,7 +10,7 @@ import java.time.ZoneId;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import static com.couponpop.security.constants.AuthTemplates.FORMATTER_YYYY_MM_DD_HH_MM_SS;
+import static com.couponpop.security.constants.SecurityTemplates.FORMATTER_YYYY_MM_DD_HH_MM_SS;
 
 @Slf4j
 @Repository
