+++ src/main/java/com/couponpop/security/constants/SecurityTemplates.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 import java.time.format.DateTimeFormatter;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class AuthTemplates {
+public final class SecurityTemplates {
 
     public static final DateTimeFormatter FORMATTER_YYYY_MM_DD_HH_MM_SS = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 }
